xylophone-main/requirements.txt
mido
pyserial
colorama
mypy
flake8
yapf
isort

